Move getColorSpaceInit to util.ts
