+++ src/webgpu/web_platform/util.ts
@@ -1,5 +1,10 @@
 import { Fixture, SkipTestCase } from '../../common/framework/fixture.js';
-import { assert, ErrorWithExtra, raceWithRejectOnTimeout } from '../../common/util/util.js';
+import {
+  assert,
+  ErrorWithExtra,
+  raceWithRejectOnTimeout,
+  unreachable,
+} from '../../common/util/util.js';
 
 declare global {
   interface HTMLMediaElement {
@@ -106,15 +111,30 @@ export function waitForNextFrame(
   return promise;
 }
 
+type VideoColorSpaceName = 'REC601' | 'REC709' | 'REC2020';
+
+export function getVideoColorSpaceInit(colorSpaceName: VideoColorSpaceName): VideoColorSpaceInit {
+  switch (colorSpaceName) {
+    case 'REC601':
+      return {
+        primaries: 'smpte170m',
+        transfer: 'smpte170m',
+        matrix: 'smpte170m',
+        fullRange: false,
+      };
+    case 'REC709':
+      return { primaries: 'bt709', transfer: 'bt709', matrix: 'bt709', fullRange: false };
+    case 'REC2020':
+      return { primaries: 'bt709', transfer: 'iec61966-2-1', matrix: 'rgb', fullRange: true };
+    default:
+      unreachable();
+  }
+}
+
 export async function getVideoFrameFromVideoElement(
   test: Fixture,
   video: HTMLVideoElement,
-  colorSpace: VideoColorSpaceInit = {
-    primaries: 'bt709',
-    transfer: 'bt709',
-    matrix: 'bt709',
-    fullRange: false,
-  }
+  colorSpace: VideoColorSpaceInit = getVideoColorSpaceInit('REC709')
 ): Promise<VideoFrame> {
   if (video.captureStream === undefined) {
     test.skip('HTMLVideoElement.captureStream is not supported');
